Don't store LDAP Bind password
Remove LDAP Bind password from the Ldap.Config interface. It is
unnecessary and undesirable to store this password in a D-Bus object.

The Ldap.Create interface still accepts the Bind password, as before. If
an admin has to change the Ldap bind password, they'd need to recreate
the LDAP config.

Change-Id: Ie5a8002bed3fd8c57500af588d9b010190f5ff9b
Signed-off-by: Deepak Kodihalli <dkodihal@in.ibm.com>